AT-SRC-001 tests: Fix float32 dtype compatibility
Fixed 4 test failures in AT-SRC-001 tests caused by dtype mismatch after
DTYPE-DEFAULT-001 migration to float32 defaults. Tests were expecting
float64 tensors but read_sourcefile() now returns float32 by default.

Changes:
- tests/test_at_src_001_simple.py: Updated dtype assertions (lines 41, 48)
- tests/test_at_src_001.py: Updated dtype assertions (lines 64, 147, 180)

Test Results: 476 passed, 117 skipped, 2 xfailed, 15 failed
- Fixed: 4 tests (AT-SRC-001 suite)
- Progress: 19 → 15 failures

Module: tests (dtype compatibility)
AT: AT-SRC-001

+++ tests/test_at_src_001.py
@@ -61,7 +61,7 @@ class TestAT_SRC_001_SourcefileAndWeighting:
 
             # Check directions are normalized (unit vectors)
             norms = torch.linalg.norm(directions, dim=1)
-            torch.testing.assert_close(norms, torch.ones(2, dtype=torch.float64))
+            torch.testing.assert_close(norms, torch.ones(2, dtype=torch.float32))
 
             # Check wavelengths match input
             assert wavelengths[0].item() == pytest.approx(1.0e-10)
@@ -144,7 +144,7 @@ class TestAT_SRC_001_SourcefileAndWeighting:
             )
 
             # Position [-15, 0, 0] normalized to unit vector: [-1, 0, 0]
-            expected_direction = torch.tensor([[-1.0, 0.0, 0.0]], dtype=torch.float64)
+            expected_direction = torch.tensor([[-1.0, 0.0, 0.0]], dtype=torch.float32)
             torch.testing.assert_close(directions, expected_direction)
 
             # Check wavelength uses default
@@ -177,7 +177,7 @@ class TestAT_SRC_001_SourcefileAndWeighting:
             torch.testing.assert_close(directions[0], -directions[1])
 
             # Both sources have weight 1.0 specified in the file
-            torch.testing.assert_close(weights, torch.ones(2, dtype=torch.float64))
+            torch.testing.assert_close(weights, torch.ones(2, dtype=torch.float32))
 
     def test_empty_sourcefile(self):
         """Test that empty sourcefile raises appropriate error."""
